+++ lib/src/annotations/interceptors.dart
@@ -1,9 +1,10 @@
 part of jaguar.src.annotations;
 
 /// Annotation to wrap one interceptor around a route
-class WrapOne {
+class WrapOne<OutputType, ResponseType, InResponseType> {
   /// Symbol of the method on the RouteClass that creates the interceptor
-  final Symbol interceptor;
+  final InterceptorCreator<OutputType, ResponseType, InResponseType>
+      interceptor;
 
   const WrapOne(this.interceptor);
 }
@@ -11,7 +12,7 @@ class WrapOne {
 /// Annotation to wrap many interceptors around a route
 class Wrap {
   /// Symbols of the methods on the RouteClass that create the interceptors
-  final List<Symbol> interceptors;
+  final List<InterceptorCreator> interceptors;
 
   const Wrap(this.interceptors);
 }